+++ src/aioway/modes/__init__.py
@@ -3,6 +3,5 @@
 from .aten import *
 from .hists import *
 from .modes import *
-from .tensors import *
 from .toggles import *
 from .tracking import *
